% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_nanopore_barcodes.R
\name{adjust_nanopore_barcodes}
\alias{adjust_nanopore_barcodes}
\title{adjust_nanopore_barcodes}
\usage{
adjust_nanopore_barcodes(
  barcodes,
  proportion_of_bases,
  volume_of_each_barcode_added = 1.25,
  same_total_volume = TRUE
)
}
\arguments{
\item{barcodes}{this is the names of the barcodes.}

\item{proportion_of_bases}{this is the proportion of base pairs that belong to each barcode. Note this should be in the same order as the names of the barcodes.}

\item{volume_of_each_barcode_added}{this is the volume that the protocol indicates each barcode to be added at. Note this should be in the same order as the names of the barcodes.}

\item{same_total_volume}{do you want the same total volume? Or do you want a different total volume? Different total volume will have the same amount of total DNA while same total volume will be even but scaled down.}
}
\value{
a tibble with intermediate calculations and final volume to be added.
}
\description{
this function is designed to take the proportion of nanopore barcodes observed in a previous run, and calculate how the user should readjust the barcodes to achieve the same number of basepairs across samples.
this is probably most usefull in the case where a flongle has been used to observe the barcode distribution of pooled sample prior to a minion run.
After seeing this distribution, the barcodes could be re-pooled to achieve an even distribution.
}
\examples{
values = runif(90)
x = values/sum(values)
barcodes = test$barcode
proportion_of_bases = test$proportion_of_bases
volume_of_each_barcode_added = 1.25
adjust_nanopore_barcodes(barcodes,proportion_of_bases,volume_of_each_barcode_added = 1.25,same_total_volume == TRUE)
}
