% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_megadb.R
\name{annotate_megadb}
\alias{annotate_megadb}
\title{annotate_megadb}
\usage{
annotate_megadb(proteinids)
}
\arguments{
\item{proteinids}{}
}
\value{
a tibble with proteinid, best_taxa_level match, description, and the database from which the annotation was derived.
note food and human are derived from uniprot.
}
\description{
annotate proteins against the megadb (hosted as a postgres sql database on AWS).
}
