% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_proteins.R
\name{annotate_proteins}
\alias{annotate_proteins}
\title{annotate_proteins
This function uses Uniprot's API to annotate proteins.}
\usage{
annotate_proteins(protein_list, columns = NULL)
}
\arguments{
\item{protein_list}{This is a list of the protein accessions that you would like to annotate}

\item{columns}{This is a string  of the column names that you would like to include followed by commas.
Column names accepted by the API can be found here: https://www.uniprot.org/help/return_fields.
If not specified, the default columns are returned.}
}
\value{
A tibble containing the protein information for the specified accessions.
}
\description{
annotate_proteins
This function uses Uniprot's API to annotate proteins.
}
