% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pubchem.R
\name{annotate_pubchem}
\alias{annotate_pubchem}
\title{annotate_pubchem}
\usage{
annotate_pubchem(input)
}
\arguments{
\item{input}{this is a list of search terms that will ultimately be  that will ultim}
}
\value{
a data frame with the input specified by user and output
}
\description{
this function allows you to search a vector containing terms against pubchem's API.
}
