% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_uniprot_single.R
\name{annotate_uniprot_single}
\alias{annotate_uniprot_single}
\title{annotate_uniprot_single
This function uses Uniprot's API to annotate proteins one protein at a time. This was made because of problems with the batch version.}
\usage{
annotate_uniprot_single(protein_list, columns = NULL)
}
\arguments{
\item{protein_list}{This is a list of the protein accessions that you would like to annotate}

\item{columns}{This is a string  of the column names that you would like to include followed by commas.
Column names accepted by the API can be found here: https://www.uniprot.org/help/return_fields.
If not specified, the default columns are returned.}
}
\value{
A tibble containing the protein information for the specified accessions.
}
\description{
annotate_uniprot_single
This function uses Uniprot's API to annotate proteins one protein at a time. This was made because of problems with the batch version.
}
\examples{
protein_list = c("P00642")
annotate_uniprot_single(protein_list)
}
