% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_plex_num.R
\name{assign_plex_num}
\alias{assign_plex_num}
\title{assign_plex_num}
\usage{
assign_plex_num(metadata, plex_grouping_factor, tmt_plex = 16)
}
\arguments{
\item{metadata}{this is a data frame containing the metadata.}

\item{plex_grouping_factor}{all samples with the same value will be assigned the same plex.
This is useful for ensuring that pared samples are in the same plex.}

\item{tmt_plex}{the n of TMT plex that you are using, either 10 or 16.}
}
\value{
a data frame with asigned plex numbers
}
\description{
this function assigns plex number for your metadata. This function also adds a bridge channel.
This is necessary for experiments that have more samples than can fit in a single plex.
}
\examples{

metadata = data.frame(sample_id = as.character(1:108),
                      plate = c(rep(1,each = 96),rep(2,each = 12)),
                      mouse_id = rep(paste0("Mouse",1:27),each = 4)
                      )

metadata_with_assigned_plex_n = assign_plex_number(metadata,"mouse_id")
}
