% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log2_p.R
\name{calc_log2_p}
\alias{calc_log2_p}
\title{calc_log2_p}
\usage{
calc_log2_p(data, column_split_by, value_name = "box_cox_scaled_values")
}
\arguments{
\item{data}{This is a data frame containing Protein ID}

\item{column_split_by}{This is the factor you would like to split data and compare by. Must only have 2 levels.}

\item{value_name}{This is the name of the value that you would like to conduct stats on.}
}
\value{
a tibble
}
\description{
This function can be used to give the log2 fold change and the pvalues from a t test of proteins between groups.
}
\examples{
# Reading and processing data
data = readr::read_delim("tests/testdata/combine_psm_fractions/PCB002_PSMs_Proteodiscover_output.txt") \%>\%
combine_psm_fractions() \%>\%
normalize_1plex() \%>\%
la_box_cox_norm() \%>\%
dplyr::mutate(Sample_ID = paste0(Sample,".",TMT))

#Reading in metadata
md = readr::read_csv("tests/testdata/metadata.csv") \%>\%
#Removed redundant columns to prevent .x and .y columns in data_md
dplyr::select(-Sample,-TMT)

# Appending md to data
data_md = dplyr::inner_join(data,md,by = "Sample_ID")

#Filtering as appropriately
f_data_md = data_md \%>\%
dplyr::filter(`Mayo_Endoscopic_Sub_Score` \%in\% c("Healthy_control","3: Severe disease (spontaneous bleeding, ulceration)")) \%>\%
dplyr::mutate_if(is.numeric,~tidyr::replace_na(.,0))

#generating stats
stats = calc_log2_p(f_data_md,"Mayo_Endoscopic_Sub_Score")
}
