% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phospho_ratio.R
\name{calc_phospho_ratio}
\alias{calc_phospho_ratio}
\title{calc_phospho_ratio}
\usage{
calc_phospho_ratio(
  proteomics_data,
  phospho_data,
  metadata,
  col_identifying_match,
  num_plexs,
  bridge_channel_plex = "126"
)
}
\arguments{
\item{proteomics_data}{this is a tibble containing proteomics data that has been read into R}

\item{phospho_data}{this is a tibble containing phosphoproteomics data that has been read into R (phospho enriched)}

\item{metadata}{this is the metadata that. Note the Plex/TMT combo needs to be concat}

\item{col_identifying_match}{this is the column in the data that identifies samples as being the same.
This is done using the metadata, and is important because this is how the ratio is actually calculated.}

\item{num_plexs}{this is the number of plexs that each dataset contains. If it is >1 things are normalized using the bridge channel}

\item{bridge_channel_plex}{this is the TMT label that contains the bridge channel to normalize to if num_plexes > 1.}
}
\value{
a tibble
}
\description{
This function calculates the ratio of phosphosite to total proteins using box cox normalized values.
}
\examples{
proteomics_data = readr::read_csv("tests/testdata/calc_phospho_ratio/pproteo_data.csv")
phospho_data = readr::read_csv("tests/testdata/calc_phospho_ratio/phospho_data.csv")
metadata = readr::read_csv("tests/testdata/calc_phospho_ratio/metadatacsv")
combined = calc_phospho_ratio("proteomics_data,phospho_data,metadata","Donor",1)
}
