% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pi_score.R
\name{calc_pi_score}
\alias{calc_pi_score}
\title{calc_pi_score}
\usage{
calc_pi_score(pval, Log2FC)
}
\arguments{
\item{pval}{: p value}

\item{Log2FC}{: log 2 fold change}
}
\value{
The pi score
}
\description{
This function returns the pi score
}
