% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_psm_fractions.R
\name{combine_psm_fractions}
\alias{combine_psm_fractions}
\title{combine_psm_fractions: Combine PSM matches from different fractions in a proteomics run.}
\usage{
combine_psm_fractions(data, analysis_type = "proteomics")
}
\arguments{
\item{data}{: This is a tibble containing the contents of the file that was exported from proteomediscoverer as a .txt file.
Note that the file names must be named following the convention of "Plex_Fraction" for this to work properly.
ie if the plex was PCB001 and the fraction number was F1, it would need to named PCB001_F1.raw.
This is because the code uses the file name to determine what plexes should be grouped together.}

\item{analysis_type}{: This is a string representing the type of analysis that is being done. Currently, proteomics and phospho are supported.}
}
\value{
a tibble containing the summed abundance for each plex.
}
\description{
This is an analogous function to the part of the TMT16/10_norm script that outputs nonnormalizeddataall.txt.
Briefly, it filters out the PSMs if their average is less than 10 across all of the plexes, if the isolation interference is > 30, or if the PSM Ambiguity == Rejected.
}
\examples{
data = read_delim("tests/testdata/normalize_to_bridge/PSM_output.txt") \%>\%
combine_psm_fractions()
}
