% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plot.R
\name{donut_plot}
\alias{donut_plot}
\title{donut_plot}
\usage{
donut_plot(data, col = "sample_type_group_3")
}
\arguments{
\item{data}{this is a tibble containing your data.}

\item{col}{this is a string containing the name of the column that you will be grouping by}
}
\value{
a ggplot2 object
}
\description{
this plots a donut plot
}
\examples{
# Plotting a donut plot
data = readr::read_csv("tests/testdata/donut_plot/donut_plot_test_data.csv")
donut_plot(data)
}
