% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sig_proteins.R
\name{extract_sig_proteins}
\alias{extract_sig_proteins}
\title{extract_sig_proteins}
\usage{
extract_sig_proteins(
  data,
  column_split_by,
  p_threshold = 0.05,
  fc_threshold = 1,
  pi_score_theshold = NULL
)
}
\arguments{
\item{data}{this is the data}

\item{column_split_by}{this is the column to split by}

\item{p_threshold}{this is the p threshold to use}

\item{fc_threshold}{this is the fc threshold to use}

\item{pi_score_theshold}{this is the pi score threshold to use if applicable}
}
\value{
a tibble
}
\description{
This function extracts the protein IDs for all proteins that have significant differences between the groups being compared.
}
\examples{
#' # Reading and processing data
data = readr::read_delim("tests/testdata/combine_psm_fractions/PCB002_PSMs_Proteodiscover_output.txt") \%>\%
combine_psm_fractions() \%>\%
normalize_1plex() \%>\%
la_box_cox_norm() \%>\%
dplyr::mutate(Sample_ID = paste0(Sample,".",TMT))

#Reading in metadata
md = readr::read_csv("tests/testdata/metadata.csv") \%>\%
#Removed redundant columns to prevent .x and .y columns in data_md
dplyr::select(-Sample,-TMT)

# Appending md to data
data_md = dplyr::inner_join(data,md,by = "Sample_ID")

#Filtering as appropriately
f_data_md = data_md \%>\%
dplyr::filter(`Mayo_Endoscopic_Sub_Score` \%in\% c("Healthy_control","3: Severe disease (spontaneous bleeding, ulceration)")) \%>\%
dplyr::mutate_if(is.numeric,~tidyr::replace_na(.,0))

#extracting the sig proteins
extract_sig_proteins(f_data_md,"Mayo_Endoscopic_Sub_Score",p_threshold = 0.05,fc_threshold = 1)
}
