% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_auc.R
\name{fraction_auc}
\alias{fraction_auc}
\title{fraction_auc
this function is designed to calculate the auc corresponding to each of the fractions from the . It is my hope that this will provide some sort of peptide quantification that can be useful when determining how much to shoot on the MS (unproven hypothesis)}
\usage{
fraction_auc(
  parsed_hplc_chromatography_data,
  start = 660,
  fraction_interval = 37
)
}
\arguments{
\item{parsed_hplc_chromatography_data}{this is the parsed hplc chromatography data, designed to be output of hplc_parser.}

\item{start}{this is the time in seconds when the method starts default 660 = 11 min}

\item{fraction_interval:}{this is the time between fraction collection intervals. Default = 37 sec.}
}
\value{
a tibble with auc calculated for each fraction.
}
\description{
Note that this function has a few method specific details hard coded into it as these don't really change in the Gonzalez lab set up. They are as follows:
}
\details{
1. A + B rows are concatenated into a single fraction for each . ie A1 + B1 get combined, A2 + B2 get combined, and so on.
2. The collector is set up to move column wise A1 -> A12 followed by B1 -> B12- then repeat.
3. The time of the method isnt evenly divisible by 37 sec, this means there is an extra amount of time where it seems like A1 gets added to at the end of the method.
}
\examples{
parsed_hplc_chromatography_data = hplc_parser("tests/testdata/hplc_parser/CB004")
fraction_auc = fraction_auc(parsed_hplc_chromatography_data)
}
