% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibson_calculator.R
\name{gibson_calculator}
\alias{gibson_calculator}
\title{gibson_calculator
This function is useful for calculating how to set up the Gibson reaction given user input lists of parts,concentrations, and lengths.}
\usage{
gibson_calculator(names, concentrations, lengths)
}
\arguments{
\item{names}{a vector of the names of each of the parts. Make sure the vector is first, followed by the inserts.}

\item{concentrations}{a vector of concentrations. Make sure that they are in the same order as names}

\item{lengths}{a vector of lengths. Make sure these are in the same order as the names}
}
\value{
a tibble containing all the ingredients at the max molarity according to the NEB protocol that is possible given the constraints imposed by the concentration of the parts. .
}
\description{
gibson_calculator
This function is useful for calculating how to set up the Gibson reaction given user input lists of parts,concentrations, and lengths.
}
\examples{
names = c("pUC19","insert_1","insert_2","insert_3")
concentrations = c(100,10,10,10)
lengths = c(5000,1000,1000,1000)

gibson_calculator(names,concentrations,lengths)
}
