% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplc_parser.R
\name{hplc_parser}
\alias{hplc_parser}
\title{hplc_parser}
\usage{
hplc_parser(directory)
}
\arguments{
\item{file_name:}{this is the file name}
}
\value{
a formated tibble with chromatography data
}
\description{
This function is designed to take the output of a .txt file that is exported from chromeleon and format it into a data frame.
To get the most out of this function, the files should be named with the following conventions:
1. an index representing the order should be placed at the beginning of the file ie 00_plex1
2. blank samples should be run between each experimental sample and should be labeled index_BLANK.
3. plexes should be names in a way that is easy to map with metadata. I use the convention PCB001 (Plex, Charlie Bayne, plex # 0001)
}
\examples{
chromatography_data = hplc_parser("tests/testdata/hplc_parser/CB004")

}
