% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_box_cox_norm.R
\name{la_box_cox_norm}
\alias{la_box_cox_norm}
\title{la_box_cox_norm
here is leigh-ana's method for conducting box_cox_normalization on proteomics data.}
\usage{
la_box_cox_norm(filepath)
}
\arguments{
\item{filepath}{: This is the filepath to a file containing the protein names in the first column and other columns containing values for each sample.}
}
\value{
a normalized csv file in the same directory.
}
\description{
la_box_cox_norm
here is leigh-ana's method for conducting box_cox_normalization on proteomics data.
}
