% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_models_per_group.R
\name{lm_models_per_group}
\alias{lm_models_per_group}
\title{lm_models_per_group}
\usage{
lm_models_per_group(data, group, yvalue, xvalue)
}
\arguments{
\item{data}{this is the data you will be operating on.}

\item{group}{this is the group you would like to do linear correlations for each member of.}

\item{yvalue}{this is the y value.}

\item{xvalue}{this is the x value.}
}
\value{
a tibble with the r squared values, p value, fdr adjusted pvalue, and direction of the correlation.
}
\description{
This function groups by a user specified variable and then conducts linear models using user specified x and y values for each group.
The p values of each linear model are then adjusted using a fdr correction.
I have found this to be useful for proteomics, and metagenomics data in the past.
In the case of proteomics, it is useful to group by protein, and then see whether the intensity of each protein correlates with an outcome such as tumor volume.
In the case of metagenomics, it can be useful to group by some level of taxonomy and then see how relative abundance correlates with an outcome such as tumor volume.
}
\examples{

data = data.frame(group = rep(1:10,each = 50), yvalue = rnorm(500,10), xvalue = rnorm(500,10))
mod = lm_models_per_group(data,"group","yvalue","xvalue")
}
