% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_1plex.R
\name{normalize_1plex}
\alias{normalize_1plex}
\title{normalize_1plex : Normalize values when only a single TMT plex is ran. This function normalizes to the median of the sample/overall median.}
\usage{
normalize_1plex(data)
}
\arguments{
\item{data}{: a tibble containing combined psms from an experiment with only one plex and no bridge chennel.}
}
\value{
a tibble containing the normalized data.
}
\description{
normalize_1plex : Normalize values when only a single TMT plex is ran. This function normalizes to the median of the sample/overall median.
}
\examples{
Here we are analyzing the results of the 1plex results.
dnorm = normalize_1plex("../")
}
