% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_bridge.R
\name{normalize_to_bridge}
\alias{normalize_to_bridge}
\title{normalize_to_bridge : Normalize PSM combine results to the bridge channel (batch correction).}
\usage{
normalize_to_bridge(data = dataframe, bridge_channel_plex = "126")
}
\arguments{
\item{data}{: This represents a data fame containing the PSMs. This is intended to be the output of combine_psm_fractions().
As such, the columns of this dataframe must follow certain naming conventions for the code to work. They are listed below.

There must be a column named Sample, this contains the name of the plex that was run.
There must also be a column named TMT, this contains the isobaric tag that was used.
There must be a column named ProteinID that has an ID that corresponds to a protein identifier.
There must be a column named value that has all of the intensity values.}

\item{bridge_channel_plex}{: This is the isobaric channel that contains the bridge channel. Defaults to 126 as that is often used in the lab.}
}
\value{
: a data frame containing the normalized results.
There are a number of different normalization steps here.

intermediate_norm = these values represent the intensity values for each sample where each value is divided by the corresponding bridge channel
value that has also been divided by the overall median of all the total bridge channel data.

value = contains the original, unnormalized value reported in the exported thermodiscover file.
final_norm = these values represent the intermediate_norm values that have been divided by the median of each sample/plex combination which have
also been divided by the median of the overall data (discounting the bridge channel)

These results can be further normalized using la_box_cox_norm(), which is leigh-ana's method for making the data even more normally distributed.
}
\description{
normalize_to_bridge : Normalize PSM combine results to the bridge channel (batch correction).
}
