% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_4MOSC1_kinetics.R
\name{parse_4MOSC1_kinetic}
\alias{parse_4MOSC1_kinetic}
\title{parse_4MOSC1_kinetics}
\usage{
parse_4MOSC1_kinetic(filepath, sheets)
}
\arguments{
\item{filepath}{this is the filepath to the completed excel sheet template
containing the data to parse}

\item{sheets}{this is a description of the sheets that you want to parse
can either be a vector of numbers or names.}
}
\value{
a tidy tibble
}
\description{
This function is designed to parse the excel files the Gutkind lab uses to
keep track of their mouse kinetic data into tidy data.
}
\details{
Will output a tibble with the following columns
1. experiment (this is the name of the tab)
2. mouse_id this is the identifier of the mice
3. cage_id: this is the experiment/cage_number combination
4. cage_number: this is the per experiment cage number (cage number is calc
by assigning first five mice to cage 1 and so on. If the total number of mice
is not divisable by 5, the last cage gets less mice).
5. measurement: the measurement taken
6. value: the value of the measurement.
7. responder_status: the assigned responder status based on final tumor volume
}
\examples{
parsed_data = parse_4MOSC1_kinetic("tests/testdata/parse_4MOSC1_kinetics/LiBIO_kinetics.xlsx",
sheets = c(2,3,6))
}
