% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cfu.R
\name{parse_cfu}
\alias{parse_cfu}
\title{parse_cfu
This function is intended to parse cfus from excel template I made for recording CFUs. The excel template can be found at
templates/parse_cfu_template.xlsx}
\usage{
parse_cfu(filepath, uL_plated = 5)
}
\arguments{
\item{filepath}{this is the file path of the filled in excel template with CFUS.}

\item{uL_plated}{this is the amount plated into spots}
}
\value{
a tibble with the parsed info.
}
\description{
parse_cfu
This function is intended to parse cfus from excel template I made for recording CFUs. The excel template can be found at
templates/parse_cfu_template.xlsx
}
\examples{
# This will parse the excel template!
cfu_counts = parse_cfu("tests/testdata/parse_cfu/colony_counts.xlsx",5)
}
