% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nanodrop_curves.R
\name{parse_nanodrop_curves}
\alias{parse_nanodrop_curves}
\title{parse_nanodrop_curves}
\usage{
parse_nanodrop_curves(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
a tibble with the following columns: sample_id, wavelength, and values.
}
\description{
This function allows you to easily read in the wavelength/intensity tsv file exported by the Nanodrop One
For some reason, the Nanodrop outputs this file in a completely unusable format.
This function parses that file and extracts the usefull information.
}
\examples{
data = parse_nanodrop_curves("tests/testdata/parse_nanodrop_cruves/nanodrop_curves.tsv")
}
