% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nanodrop_summary.R
\name{parse_nanodrop_summary}
\alias{parse_nanodrop_summary}
\title{parse_nanodrop_summary}
\usage{
parse_nanodrop_summary(filepath)
}
\arguments{
\item{filepath}{this is the file path to the csv that was exported from the nandrop.}
}
\value{
a tibble with the first 7 columns (only important ones) from the file.
}
\description{
This function allows you to easily read in the summary csv file exported by the nanodrop one
For some reason, the Nanodrop outputs the summary csv file in a weird format.
It is actually tab separated (instead of comma) and is UTF-16 encoded.
This is all stuff I tend to forget, so here is a function to easily read this data in.
}
\examples{
data = parse_nanodrop_summary("tests/testdata/parse_nanodrop_summary/nanodrop_summary.csv")
}
