% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_stratus.R
\name{parse_stratus}
\alias{parse_stratus}
\title{parse_stratus}
\usage{
parse_stratus(input, wavelength_nm = 600)
}
\arguments{
\item{input}{this is the .csv file exported from the stratus plate reader.}
}
\value{
a tibble with the following columns.
time_hr -> the amount of time in hrs that have elapsed since the start of the read
datetime -> the date time of the measurement.
temperature -> the temperature that was recorded at the time of the measurement in C
well -> the well corresponding to a measurement.
paste0("od",wavelength) -> values of the measured wavelength. Our machine measured 600nm wavelength so default is od600.
}
\description{
This function parses the output from the stratus plate reader by cerillo (https://cerillo.bio/stratus/) into a usable form.
It takes the  csv file exported by the plate reader and transforms it into a tidy data format.
}
\examples{
parsed_OD600_data = parse_stratus("tests/testdata/parse_stratus/stratus_data_export.csv")
}
