% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_quant.R
\name{peptide_quant}
\alias{peptide_quant}
\title{peptide_quant}
\usage{
peptide_quant(filepath)
}
\arguments{
\item{filepath}{this is the filepath to an exported .txt file from either the peptide_quant_od480_single_plate.spr or peptide_quant_od480_single_plate.spr protocols.
Note that when doing the export for this file you need to only check the standards and unknowns boxes.}
}
\value{
a tibble containing processed data. Note that predicted values are calculated for unknown sample concentrations using a linear model.
}
\description{
this function takes the export from the plate reader and formats data into a tidy format while calculating predicted concentrations of unknowns using a linear model of the standard curve.
}
\examples{
# Single plate
single = peptide_quant("tests/testdata/peptide_quant/single_plate_od480.txt")
multi = peptide_quant("tests/testdata/peptide_quant/multi_plate_od480.txt")
}
