% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_venn_diagram.R
\name{phospho_venn_diagram}
\alias{phospho_venn_diagram}
\title{phospho_venn_diagram}
\usage{
phospho_venn_diagram(
  proteomics_data,
  phospho_enriched_data,
  bridge_channel = NULL
)
}
\arguments{
\item{proteomics_data}{output from a proteome discoverer exported text file acquured from a non phospho enriched experiment. Intended to have been read into R using read_delim().
Note that the workflow used in proteome discoverer must include the ptmRS node and dynamic phospho mods in Sequest.}

\item{phospho_enriched_data}{from a paired phospho-enriched experient (same samples as proteomics data). Intended to have beenread into R using read_delim().
Note that the workflow used must include the ptmRS node and dynamic phospho mods in Sequest.}

\item{bridge_channel}{this is the bridge channel used if applicable. Default is NULL meaning that data will be processed as a single plex unless otherwise indicated.}
}
\value{
a ggplot2 object
}
\description{
This function takes the results from a non-phospho enriched proteomics experiment (Sequest with dynamic phospho mode and ptmRS node enabled), looks at uniwue Protein-Sequence-PTM combinations
and compares them to the results from a phospho enriched proteomics experiment (Sequest with dynamic phospho mode and ptmRS node enabled).
}
