% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_phospho_mod.R
\name{psm_phospho_mod}
\alias{psm_phospho_mod}
\title{psm_phospho_mod
This code modifies the PSM file that is output from proteome discoverer to facilitate phospho PTM analysis.}
\usage{
psm_phospho_mod(data)
}
\arguments{
\item{data}{This is the=PSM proteome discoverer output file that has been into R as a tibble}
}
\value{
a tibble with a Protein ID column that is compatible with downstream processing through the combine_psm_fractions() and other functions.

Note that the concatination of Master Accession, Annotated Sequence and ptmRS are concatinated into the ProteinID column.
Even though the name ProteinID isn't named exactly correctly, it was much easier to do it this way since that was the convention susequent code uses.
}
\description{
psm_phospho_mod
This code modifies the PSM file that is output from proteome discoverer to facilitate phospho PTM analysis.
}
