% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_url_request.R
\name{pubchem_url_request}
\alias{pubchem_url_request}
\title{pubchem_url_request}
\usage{
pubchem_url_request(
  input,
  search1 = "compound",
  search2 = "smiles",
  operation = "synonyms"
)
}
\arguments{
\item{input}{this is the input}

\item{search1}{this is the first search string}

\item{search2}{this is the second search}

\item{operation}{this is the operation that you would like to perform on the data}
}
\value{
a url that hits pubchem's api and can be used to search for user specified data.
}
\description{
This function generates a pubchem url request for user specified info. designed to be used in combination with the
}
\examples{
pub_chem_smiles = "CN1C2CCC1C(C(C2)OC(=O)C3=CC=CC=C3)C(=O)OC"
pubchem_url_request(pub_chem_smiles)
}
