% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_plates.R
\name{randomize_plates}
\alias{randomize_plates}
\title{randomize_plates}
\usage{
randomize_plates(metadata)
}
\arguments{
\item{metadata}{this is the metadata for your experiment.}
}
\value{
metadata dataframe with randomized plate assignment per sample.
}
\description{
this function will randomize 96 well plate location given a dataframe containing metadata.
Note that this will fill as many plates as can be completely and then add samples to the last (n < 96) remaining plate randomly.
}
\examples{

metadata = data.frame(sample_id = 1:108)
metadata_with_randomly_assigned_plates= randomize_plates(metadata)

}
