% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_tmt.R
\name{randomize_tmt}
\alias{randomize_tmt}
\title{randomize_tmt}
\usage{
randomize_tmt(metadata, tmt_plex = 16)
}
\arguments{
\item{metadata}{this is a data frame containing at the minimum a column named plex_number.
This column is
Note that this should include the bridge channel for each plex.}

\item{tmt_plex}{this specifies whether a 10 or 16 plex is being used.}
}
\value{
a data frame with randomly assigned plex numbers
}
\description{
This function will provide randomized tmt labels for given plex groups.
}
\examples{

metadata = data.frame(sample_id = 1:30,plex_num = rep(c(1,2), each = 15))
metadata_with_randomly_assigned_tmts = randomize_tmt(metadata)

}
