% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_wells.R
\name{randomize_wells}
\alias{randomize_wells}
\title{randomize_wells}
\usage{
randomize_wells(metadata)
}
\arguments{
\item{metadata}{this is a dataframe containing the metadata.
Note that it must have a column titled plate which contains the plate number of each sample.}
}
\value{
a tibble with randomized wells.
}
\description{
This function will randomize well location given a data frame of metadata.
}
\examples{
metadata = data.frame(sample_id = 1:108,plate = c(rep(1,each = 96),rep(2,each = 12)))
metadata_with_randomly_assigned_wells= randomize_wells(metadata)
}
