% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_compact_peak_table.R
\name{read_compact_peak_table}
\alias{read_compact_peak_table}
\title{read_compact_peak_table
easily read in data from compact peak table export from Tapestation.}
\usage{
read_compact_peak_table(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
a tibble that has been processed to remove
}
\description{
read_compact_peak_table
easily read in data from compact peak table export from Tapestation.
}
\examples{
data = read_compact_peak_table("tests/read_compact_peak_table/compactPeakTable.csv")
}
