% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_well_col.R
\name{split_well_col}
\alias{split_well_col}
\title{split_well_col
This function will take a well coordinate (formatted like A01) and return the col ie 01.}
\usage{
split_well_col(well)
}
\arguments{
\item{well}{this is a vector of well coordinates (Specified as LetterNumber)}
}
\value{
a character vector
}
\description{
split_well_col
This function will take a well coordinate (formatted like A01) and return the col ie 01.
}
\examples{
wells = c("A1","A2")
cols = split_well_col(wells)
}
