% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_well_row.R
\name{split_well_row}
\alias{split_well_row}
\title{split_well_row
This function will take a well coordinate (formatted like A01) and return the row ie A.}
\usage{
split_well_row(well)
}
\arguments{
\item{well}{this is a vector of well coordinates (Specified as LetterNumber)}
}
\value{
a character vector
}
\description{
split_well_row
This function will take a well coordinate (formatted like A01) and return the row ie A.
}
\examples{
wells = c("A1","A2")
rows = split_well_row(wells)
}
