% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_log2.R
\name{t_test_log2}
\alias{t_test_log2}
\title{t_test_log2}
\usage{
t_test_log2(data, column_split_by)
}
\arguments{
\item{data}{: This is the inbut data. Intended to be the output of la_box_cox_norm()}

\item{column_split_by}{:  this is the name of the column you would like to split data by in order to compare groups. Should only contain 2 levels.
Note that it must contain only 2 groups, and the column name cannot have any spaces}
}
\description{
This function reports the log2 change, t test p values/ p.adj values when comparing two groups.
}
