% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_food.R
\name{annotate_food}
\alias{annotate_food}
\title{annotate_food
This function can used to annotate foods into higher groupings.
ie a pig can be classified as red meat, animal, etc depending on the level of granularity.}
\usage{
annotate_food(common_names)
}
\arguments{
\item{common_names}{a list of the common names in a database}
}
\value{
a tibble  containing
}
\description{
annotate_food
This function can used to annotate foods into higher groupings.
ie a pig can be classified as red meat, animal, etc depending on the level of granularity.
}
\examples{
#list of protein IDs
proteinids = read_csv("tests/testdata/megadb/proteinids.csv") \%>\%
 pull(ProteinID)

megadb_results = annotate_megadb(proteinids)

#pulling common names
common_names = megadb_results \%>\%
dplyr::pull(common_name) \%>\%
dplyr::unique()

#Now we can figure out what these foods are
foods = annotate_food(common_names)
}
