% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plot.R
\name{donut_plot}
\alias{donut_plot}
\title{donut_plot}
\usage{
donut_plot(col)
}
\arguments{
\item{col}{this is a string containing the name of the column that you will be grouping ny}
}
\value{
a ggplot2 donut plot
}
\description{
donut_plot
}
