% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_box_cox_norm.R
\name{la_box_cox_norm}
\alias{la_box_cox_norm}
\title{la_box_cox_norm
here is leigh-ana's method for conducting box_cox_normalization on proteomics data.}
\usage{
la_box_cox_norm(data)
}
\arguments{
\item{data}{: This is a tibble containing at least the columns named Sample , TMT, and final_norm. This is intended to be the output of normalize_to_bridge() or noramlize_1plex()}
}
\value{
a normalized csv file in the same directory.
}
\description{
la_box_cox_norm
here is leigh-ana's method for conducting box_cox_normalization on proteomics data.
}
