% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_1plex.R
\name{normalize_1plex}
\alias{normalize_1plex}
\title{normalize_1plex : Normalize values when only a single TMT plex is ran. This function normalizes to the median of the sample/overall median.}
\usage{
normalize_1plex(data, data_format = "long")
}
\arguments{
\item{data}{: a tibble containing combined psms from an experiment with only one plex and no bridge chennel.}

\item{data_format}{: This is a character string specifying whether you would like the data in long or wide format. Note that if data_format ==  wide only data
corresponding to final_norm is returned. The intermediate normalization steps are disregarded in this case.}
}
\value{
a tibble containing the normalized data.
}
\description{
normalize_1plex : Normalize values when only a single TMT plex is ran. This function normalizes to the median of the sample/overall median.
}
\examples{
data = read_delim("tests/testdata/combine_psm_fractions/PCB002_PSMs_Proteodiscover_output.txt") \%>\%
combine_psm_fractions() \%>\%
normalize_1plex()
}
