% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot
this plots a volcano plot with proteins that meet the specified criteria highlighted.}
\usage{
volcano_plot(data, column_split_by, p_threshold = 0.05, fc_threshold = 1)
}
\arguments{
\item{data}{= this is a tibble that has been loaded into R.}

\item{column_split_by}{= this is the name of the column you would like to split data by in order to compare groups
Note that it must contain only 2 groups, and the column name cannot have any spaces.}

\item{p_threshold}{= this is the adjusted p value threshold that is used to generate the dashed line on the volcano plot}

\item{fc_threshold}{= this is the fc_threshold that is used to generate the dashed line on the volcano plot.}

\item{pi_score_theshold}{= if used, the indicated pi score threshold is used instead of the p_threshold and fc_threshold.}
}
\value{
a ggplot2 object
}
\description{
volcano_plot
this plots a volcano plot with proteins that meet the specified criteria highlighted.
}
\examples{
# Reading and processing data
data = readr::read_delim("tests/testdata/combine_psm_fractions/PCB002_PSMs_Proteodiscover_output.txt") \%>\%
combine_psm_fractions() \%>\%
normalize_1plex() \%>\%
la_box_cox_norm() \%>\%
dplyr::mutate(Sample_ID = paste0(Sample,".",TMT))

#Reading in metadata
md = readr::read_csv("tests/testdata/metadata.csv") \%>\%
#Removed redundant columns to prevent .x and .y columns in data_md
dplyr::select(-Sample,-TMT)

# Appending md to data
data_md = dplyr::inner_join(data,md,by = "Sample_ID")

#Filtering as appropriately
f_data_md = data_md \%>\%
dplyr::filter(`Mayo_Endoscopic_Sub_Score` \%in\% c("Healthy_control","3: Severe disease (spontaneous bleeding, ulceration)")) \%>\%
dplyr::mutate_if(is.numeric,~tidyr::replace_na(.,0))

#Volcano plot
volcano_plot(f_data_md,"Mayo_Endoscopic_Sub_Score",p_threshold = 0.05,fc_threshold = 1)
}
